requirements.txt
streamlit
pandas
librosa
tsfel
scikit-learn
numpy
joblib
streamlit-audiorecorder